Feat updateimg

See merge request HHP/sysroot_docker!5

+++ usr/include/keros_i2c_interface.h
@@ -19,7 +19,7 @@
 struct keros_i2c_msg {
     unsigned short addr;
     unsigned short flags;
-    unsigned short len;
+    int len;
     unsigned char *buf;
 };
